requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
pydantic==2.8.2
python-multipart==0.0.9
Pillow==10.4.0
numpy==1.26.4
pandas==2.2.2
requests==2.32.3
streamlit==1.38.0
pytest==8.3.2